Initial implementation for SBSpergel and galsim.Spergel classes
This implementation successfully compiles and instantiates a `galsim.Spergel(nu)` object,
with index `nu`. The implementation is copied from SBSersic and SBExponential, with
most optimizations removed. Also see the 'TODO' items in SBSpergel.cpp for required
functionality that is not yet implemented properly.

When copying the SBSersic implementation, I removed all the functionality for truncated
profiles. But truncation may be required here because the Spergel profile is similar to the
Sersic profile. It seemed better to build up the necessary features rather than copy
potentially unnecessary cruft.