add a new profile RandomWalk
This class generates a set of point sources.  The location of these
is based on a random walk with specified number of steps.

the user can choose the number of point sources, number of steps,
the expected half-light radius of the distribution, and the total
flux in the points.

The points are represented as small galsim.Gaussian objects