Re-enable/improve test_structure_function
Previous version only tested statistics of a Kolmogorov screen.  It
turns out that the Kolmogorov power spectrum is the hardest one to test,
though, which I suspect is related to the diverging phase power as
k -> 0, and so the tolerance had been set pretty high at 50%.  Now that
VonKarman exists, we can test simulated screens against more complicated
analytic structure functions, which allows the tolerance to go down to
5%.  This is particularly important given the sqrt(2) anomaly in the
previous version of the phase screen generation code.