Use photo shooting for point sources in RandomWalk
A profile is now used to generate the points with photon shooting

One can send the profile= keyword to specify the desired profile.  In
that case, the hlr and flux are taken from the profile

The old method, sending hlr and flux explicitly, is still supported and
internally generates a Gaussian profile.

The __repr__ and __str__ were updated to always use the profile for
simplicity