Add a spiral galaxy class
The spiral is a lograthmic spiral in 3d that is projected onto 2d.  The
profile is represented by a number of point sources.  The number
of arms, inclination, rotation and hlr/disk height are configurable,
as well as the pitch angle of the arms.

Unit tests are also added in tests/test_spiral.py